Fetch remote `dylint_driver` in `dylint_testing` by default
The driver builder uses a remote `dylint_driver` package if building
in release mode or if `dylint_driver_remote` is enabled.

https://github.com/smoelius/dylint/blob/844a3ab339d0dcf69102044368b1bbd907b45787/dylint/src/driver_builder.rs#L112-L113

So when `dylint_testing` is built in debug mode, it looks for a local
`dylint_driver` package and cannot find one.

A solution is to have `dylint_testing` enable the `dylint_driver_remote`
feature by default.

I think I would prefer the feature be `dylint_driver_local`, but this
solves the immediate problem.